+++ src/main/java/br/com/rogerio/Dollar.java
@@ -1,8 +1,6 @@
 package br.com.rogerio;
 
-public class Dollar {
-
-	private int amount;
+public class Dollar extends Money{
 
 	public Dollar(int amount) {
 		this.amount = amount;
@@ -12,25 +10,4 @@ public class Dollar {
 		return new Dollar(amount * multiplier);
 	}
 
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + amount;
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		Dollar other = (Dollar) obj;
-		if (amount != other.amount)
-			return false;
-		return true;
-	}
 }