+++ src/main/java/br/com/rogerio/Franc.java
@@ -1,8 +1,6 @@
 package br.com.rogerio;
 
-public class Franc {
-
-	private int amount;
+public class Franc extends Money {
 
 	public Franc(int amount) {
 		this.amount = amount;
@@ -10,27 +8,5 @@ public class Franc {
 
 	public Franc times(int multiplier) {
 		return new Franc(amount * multiplier);
-	}
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + amount;
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		Franc other = (Franc) obj;
-		if (amount != other.amount)
-			return false;
-		return true;
 	}
 }